Move makeBatch and getChunkSize to their own file, batching.ts
