+++ ts/packages/agents/spelunker/src/embeddings.ts
@@ -12,7 +12,7 @@ import { NormalizedEmbedding } from "typeagent";
 import { Chunk, ChunkId } from "./chunkSchema.js";
 import { console_log } from "./logging.js";
 import { retryOn429 } from "./retryLogic.js";
-import { makeBatches } from "./searchCode.js";
+import { makeBatches } from "./batching.js";
 import { SpelunkerContext } from "./spelunkerActionHandler.js";
 
 export function makeEmbeddingModel(): TextEmbeddingModel {
